Fix commands
