+++ example/erc20/enclave/src/ecalls.rs
@@ -13,15 +13,21 @@ register_ecall!(
     MAX_MEM_SIZE,
     Runtime<AnonifyEnclaveContext>,
     AnonifyEnclaveContext,
-    (SEND_COMMAND_CMD, CmdSender<Ed25519ChallengeResponse>),
+    (
+        SEND_COMMAND_TREEKEM_CMD,
+        CommandByTreeKemSender<Ed25519ChallengeResponse>
+    ),
     // Fetch a ciphertext in event logs from blockchain nodes into enclave's memory database.
-    (FETCH_CIPHERTEXT_CMD, CmdReceiver<Ed25519ChallengeResponse>),
+    (
+        FETCH_CIPHERTEXT_TREEKEM_CMD,
+        CommandByTreeKemReceiver<Ed25519ChallengeResponse>
+    ),
     // Fetch handshake received from blockchain nodes into enclave.
-    (FETCH_HANDSHAKE_CMD, HandshakeReceiver),
+    (FETCH_HANDSHAKE_TREEKEM_CMD, HandshakeReceiver),
     // Get current state of the user represented the given public key from enclave memory database.
     (GET_STATE_CMD, GetState<Ed25519ChallengeResponse>),
-    (JOIN_GROUP_CMD, JoinGroupSender),
-    (SEND_HANDSHAKE_CMD, HandshakeSender),
+    (JOIN_GROUP_TREEKEM_CMD, JoinGroupSender),
+    (SEND_HANDSHAKE_TREEKEM_CMD, HandshakeSender),
     (
         REGISTER_NOTIFICATION_CMD,
         RegisterNotification<Ed25519ChallengeResponse>
@@ -40,15 +46,21 @@ register_ecall!(
     MAX_MEM_SIZE,
     Runtime<AnonifyEnclaveContext>,
     AnonifyEnclaveContext,
-    (SEND_COMMAND_CMD, CmdSender<Ed25519ChallengeResponse>),
+    (
+        SEND_COMMAND_TREEKEM_CMD,
+        CommandByTreeKemSender<Ed25519ChallengeResponse>
+    ),
     // Fetch a ciphertext in event logs from blockchain nodes into enclave's memory database.
-    (FETCH_CIPHERTEXT_CMD, CmdReceiver<Ed25519ChallengeResponse>),
+    (
+        FETCH_CIPHERTEXT_TREEKEM_CMD,
+        CommandByTreeKemReceiver<Ed25519ChallengeResponse>
+    ),
     // Fetch handshake received from blockchain nodes into enclave.
-    (FETCH_HANDSHAKE_CMD, HandshakeReceiver),
+    (FETCH_HANDSHAKE_TREEKEM_CMD, HandshakeReceiver),
     // Get current state of the user represented the given public key from enclave memory database.
     (GET_STATE_CMD, GetState<Ed25519ChallengeResponse>),
-    (JOIN_GROUP_CMD, JoinGroupSender),
-    (SEND_HANDSHAKE_CMD, HandshakeSender),
+    (JOIN_GROUP_TREEKEM_CMD, JoinGroupSender),
+    (SEND_HANDSHAKE_TREEKEM_CMD, HandshakeSender),
     (
         REGISTER_NOTIFICATION_CMD,
         RegisterNotification<Ed25519ChallengeResponse>
