+++ frame/common/src/ciphertexts/mod.rs
@@ -1,1 +1,2 @@
+pub mod enclave_key;
 pub mod treekem;
