+++ modules/anonify-enclave/src/commands/mod.rs
@@ -1,0 +1,4 @@
+pub mod executor;
+pub mod plaintext;
+pub mod treekem;
+pub mod enclave_key;
